Use standard Apache-2.0 license file
This uses the standard Apache-2.0 license file, available for
download from https://www.apache.org/licenses/LICENSE-2.0.txt, as
`LICENSE-APACHE`. The license text itself is unchanged, but this
fixes the broken license appendix. The appendix is expressly not
part of the licese terms, so probably nothing very bad would happen
due to it, but it is better either to have the appendix, or to omit
it entirely, than to have only some fragments of it.

At the very beginning, gitoxide was licensed under only the MIT
license. Early on, all contributors agreed to dual-license gitoxide
under the MIT license and the Apache-2.0 license (#8), and an
Apache-2.0 license file was added (ea353eb).

The standard Apache-2.0 license file ends in an appendix that
describes the usual way of explicitly applying it to code, which
contains placeholders which are meant to be substituted if the
boilerplate code is copied elsewhere, but not in the license file
appendix itself. However, in ea353eb, the placeholders were
substituted in the appendix in the license file itself, and the
more instructional portion of the appendix was removed.

This modification to the appendix, which created an unusual license
file, may have been done in order to put the copyright notice
somewhere where it would be specifically associated with the
Apache-2.0 license option. After all, the boilerplate text wasn't
(and intentionally continues not to be) used as a header in source
code files as the appendix suggests. But this carried two problems:

1. It was potentially confusing with respect to the significance of
   that text, since it was not present anywhere a copyright notice
   would be expected, and came after "END OF TERMS AND CONDITIONS".

2. It had the potential to confuse tooling that processed licenses.

One specific case of (2) is known, described in 76ae5d6 (#1232)
where the license file was changed to remove it. (The MIT license
file was also changed to remove the copyright notice, but it is
unlikely that the MIT license file contributed to tooling problems,
since the copyright line is expected in an MIT license file and is
typically present.)

That left fragments of the appendix in the Apache-2.0 license file,
which no longer even attempts to give any information more specific
than that which is present in the licese terms themselves.

There are two good ways to fix the problem. One of them is to use
the standard version of the Apache-2.0 license file, with the full
original appendix with instructions and unsubstituted placeholders
intact. That approach is followed here, since adding the missing
pieces of the standard appendix makes clear what the current
nonstandard fragment is from.

(The other approach is to remove the appendix altogether, which is
fine to do since it is expressly not part of the license terms, and
which seems to be fairly popularly done among Rust projects. We may
end up going with that, but for clarity it's not done just yet.)