Further upgrade `jiff` to fix fuzz failures (#1984)
This upgrades `jiff` to 0.2.12 to fix fuzzing failures. The test
case introduced in the previous commit now passes.

Upgrading `jiff` past 0.2.11 gets the fix for BurntSushi/jiff#359,
fixed in BurntSushi/jiff#360. To avoid declaring compatibility with
a version that has the bug, this advances `jiff` from 0.2.10 to
0.2.12 in the `gix-archive` and `gix-date` manifests.

(The changes in `Cargo.lock` are those that occur automatically
when `cargo check --workspace` is run after those changes.)

Fixes #1979
Fixes #1982