Use `-p` where no longer ambiguous
The previous commit notes:

> Before this, some actions we'd prefer to do by `<cmd> -p <crate>`
> had to be done by `(cd <crate-dir>; <cmd>)`. This was needed to
> operate on `gix-*` crates in the workspace that are also
> dependencies, even transitively, of `gix-testtools`.

All occurrences of `cd ...` followed immediately by a `cargo`
command in the `justfile` and in `ci.yml` were for that reason.

This commit changes them to use `-p` instead. The changes in the
`justfile` affect both manual runs and runs on CI through the
`test` job. The changes directly in `ci.yml` affect the `wasm`
jobs.

This is intended to be a refactoring. No change is anticipated in
what tests are run, their features, or their behavior.

The rationale is twofold:

- Simplify the commands. This is the form most other such commmands
  were already written in. These seem to have been written in the
  `cd` form only as a workaround for the now-fixed `-p` ambiguity.

- Verify that the fix actually works and that there is nothing else
  breaking these checks when they are run from the top level of the
  workspace.