Rewrite expanded `ci-check-msrv` doc so `just --list` works
This rewrites the newly expanded multi-line comment explaining what
the `ci-check-msrv` recipe in the `justfile` does, so that it is
still readable from top to bottom, but so that the one-line summary
is the last line.

This is needed because `just --list` only treats a single `#` line,
preceding the formal beginning of the recipe, as its documentation.

(This differs from most other uses of multi-line comments, such as
in a Rust `///` or `//!` comment, where the first line could be a
short summary, and subsequent paragraphs are regarded subordinate.)

This continues to change only the `justfile` and not to modify the
`##` comment in the corresponding `Makefile` rule.