Use `--locked` instead of checking `Cargo.lock` afterwards
The verification being done here is exactly what `--locked` is for.

Allowing the check to proceed even if it changes the dependencies
could have the benefit of showing more information. But it is not
obvious that the information it shows would be relevant to the
situation being investigated.

(If that information would usually be relevant, then
`cargo +nightly update -Zminimal-versions` may be too strong, and
`cargo +nightly update -Zdirect-minimal-versions` could be
considered. Even better than either of those *might* be, somehow,
to temporarily downgrade locked dependencies only as much as
necessary to make them all MSRV-compatible.)

In any case, `--locked` is more readable, and considerably simpler.

In addition to making complementary changes to the `msrv.yml`
workflow and to the `justfile` recipe it uses, this also changes
the `Makefile` rule corresponding to the `justfile` recipe, so they
continue to do the same thing if used in local experiments.