Further split `check-msrv` steps; let Windows use `pwsh` again
This commit splits the commands in the steps of the `check-msrv`
job definition to be one per step, documenting each step.

In 510847b I predicted that it would not make sense to split the
steps to one command per step in `msrv.yml`. The change made there
of using `bash` even on Windows allowed for experimenting with how
the workflow logs look when reorganized in various ways. But in
view of other changes -- and to better clarify that two toolchains
were installed, but the MSRV toolchain is set default, so it is
used by all subsequent operations *except* the `-Zminimal-versions`
dependency downgrade that currently requires `nightly` -- it now
looks like having one command per step is better after all.

When running only one (simple) command per step, the main change in
510847b is no longer needed. That is, this now avoids that problem
in the same way it has been avoided in `test-fast` in `ci.yml`
since 4f2ab5b (#1559). So this commit also removes `shell: bash` in
`msrv.yml` (but keeps the comment clarification in `ci.yml`).