Actually build `gix` to check MSRV
This changes `cargo check` commands to `cargo build` in the
`ci-check-msrv` recipe in the `justfile`, which the `check-msrv` CI
jobs in `msrv.yml` use. (It updates a CI step name accordingly.)

The idea is to make sure at least `gix`, with the two combinations
of features tested, actually *builds* under the MSRV toolchain.

As in f10f18d, this also updates the `Makefile` rule corresponding
to that `justfile` recipe.

The idea of actually building was suggested in:
https://github.com/GitoxideLabs/gitoxide/issues/1808#issuecomment-2614260268

However, this does not uncover any new breakages. And there has
been further improvement on #1808, including in the commits leading
up to this, as well as earlier, in 569c186 (#1909). Nonetheless, it
seems likely that some problems remain with some combinations of
crates and features that are not currently exercised in the MSRV
check. #1808 is most likely not yet fully fixed.